split typography into heading and text components
