+++ src/Heading/index.ts
@@ -1,4 +1,2 @@
 export * from './Heading';
 export { default as Heading } from './Heading';
-export * from './Text';
-export { default as Text } from './Text';
