+++ src/Text/index.ts
@@ -1,0 +1,2 @@
+export * from './Text';
+export { default as Text } from './Text';
