+++ src/index.tsx
@@ -1,4 +1,5 @@
 export * from './Box';
 export * from './Button';
+export * from './Heading';
+export * from './Text';
 export * from './Theme';
-export * from './Typography';
